+++ methodical-core/src/test/java/org/jwcarman/methodical/DefaultMethodInvokerAccessibilityTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jwcarman.methodical.def;
+package org.jwcarman.methodical;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -22,7 +22,6 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import org.jwcarman.methodical.MethodInvoker;
 import org.jwcarman.methodical.param.ParameterResolver;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)